trading_platform/requirements.txt
# API Requirements
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5
numpy>=1.20.0
python-dotenv>=1.0.0

# Database (optional)
psycopg2-binary>=2.9.0

# GUI Requirements
PyQt5>=5.15.0
pandas>=1.5.0
requests>=2.28.0

# Shared Requirements
python-dateutil>=2.8.0